Load from GitCDN rather than githack
While the cache update time is slower on GitCDN (a couple hours vs a
couple minutes), the service is designed to allow you to always pull the
latest commit, rather than needing to pull a specific version in
production. This allows us to not have to change loader.js when hifoc.js
is changed without putting unnecessary load on a free server.